docu and stuff

+++ tools/clean_device_signals.py
@@ -9,9 +9,11 @@ import dash
 from dash.exceptions import PreventUpdate
 from pyadlml.constants import ACTIVITY, DEVICE, END_TIME, START_TIME, STRFTIME_PRECISE, TIME, VALUE
 from pyadlml.dataset._core.activities import read_activity_csv
-from pyadlml.dataset._core.devices import _generate_signal, device_remove_state_matching_signal, read_device_df, write_device_df
+from pyadlml.dataset._core.devices import _generate_signal, create_sig_and_corr, device_remove_state_matching_signal, read_device_df, write_device_df
+from pyadlml.dataset._datasets.activity_assistant import write_device_map, write_devices
 from pyadlml.dataset.plot.plotly.acts_and_devs import activities_and_devices
 from pyadlml.dataset.plot.plotly.util import dash_get_trigger_element, dash_get_trigger_value, remove_whitespace_around_fig
+from pyadlml.constants import ts2str
 
 from pyadlml.dataset.util import df_difference, fetch_by_name
 from pyadlml.dataset.plot.plotly.activities import correction
@@ -36,24 +38,6 @@ for dataset options look under pyadlml.constants:
 """
 
 SEP_HIT = '#' + '-'*50
-
-def create_sig_and_corr(df: pd.DataFrame, hit_idx: int, dt_prae_hit: pd.Timedelta, dt_post_hit: pd.Timedelta):
-    """ 
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        
-
-    """
-    idx, first, last = get_idxs(df, hit_idx, dt_prae_hit, dt_post_hit)
-    tmp = df.loc[idx, [VALUE, 'diff']].values
-    tmp = np.insert(tmp, 0, [[not tmp[0, 0], first]], axis=0)
-    tmp = np.append(tmp, [[not tmp[-1, 0], last]], axis=0)
-    signal = _generate_signal(tmp, dt='250ms')
-    corr = sc_signal.correlate(signal, ss_discrete, mode='full')
-    is_match = auto_corr_max - corr.max() < eps_corr
-    return signal, corr, is_match
 
 @remove_whitespace_around_fig
 def plot_cc(sig_match, sig_search, auto_corr, cross_corr, t, eps_corr, cc_max = None):
@@ -126,66 +110,8 @@ def plot_cc(sig_match, sig_search, auto_corr, cross_corr, t, eps_corr, cc_max = 
 
 
 
-def get_idxs(df, hit_idx, dt_prae_hit, dt_post_hit):
-    """ Gets device indices of events before and after to the hit 
-        to later on process in a cross correlation.
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        device dataframe of the selected device
-    hit_idx: int
-        Indice of the hit
-    dt_prae_hit: pd.Timedelta
-        How much time to consider prea hit
-    dt_prae_hit: pd.Timedelta
-        How much time to consider post hit
-
-    Returns
-    -------
-    list, 
-        The involved indices for the signal
-    """
-    get_df_idx = lambda x: df[df['index'] == x].index.values[0]
-
-    res_idxs = [hit_idx]
-    df = df.reset_index()
-
-    if get_df_idx(hit_idx) == 0:
-        first = dt_prae_hit
-    else:
-        # Iterate the events backward and finish if iterated time difference
-        # is higher than the origs. signals 
-        idx_lower = get_df_idx(hit_idx) - 1
-        dt_iter = dt_prae_hit
-        dt_prev_tmp = dt_iter - df.at[idx_lower, 'diff']
-        while dt_prev_tmp > pd.Timedelta('0s') and idx_lower >= 0:
-            dt_iter -= df.at[idx_lower, 'diff']
-            res_idxs.insert(0, df.at[idx_lower, 'index'])
-            idx_lower -= 1
-            dt_prev_tmp -= df.at[idx_lower, 'diff']
-        first = dt_iter
-
-    if get_df_idx(hit_idx) == len(df) - 1:
-        last = dt_post_hit
-    else:
-        # Iterate the events forward and finish if iterated time difference
-        # is higher than the origs. signals 
-        idx_upper = get_df_idx(hit_idx) + 1
-        dt_iter = dt_post_hit
-        dt_next_tmp = dt_iter - df.at[idx_upper, 'diff']
-        while dt_next_tmp > pd.Timedelta('0s') and idx_upper < len(df) - 1:
-            dt_iter -= df.at[idx_upper, 'diff']
-            res_idxs.append(df.at[idx_upper, 'index'])
-            idx_upper += 1
-            dt_next_tmp -= df.at[idx_upper, 'diff']
-        last = dt_iter
-
-    df = df.set_index('index')
-    return res_idxs, first, last
-
 def del_op2str(row: pd.Series):
-    row_lst = f"\t[['{row[TIME]}','{row[DEVICE]}','{row[VALUE]}']]"
+    row_lst = f"\t[['{ts2str(row[TIME])}','{row[DEVICE]}','{row[VALUE]}']]"
     return f"idx_to_del = get_index_matching_rows(df_devs, {row_lst})\n"\
             + f"df_devs = df_devs.drop(index=idx_to_del).reset_index(drop=True)\n\n"
 
@@ -215,7 +141,7 @@ def plot_main(df_devs, df_acts, dev_name, signal_orig, zoomed_range):
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--device', type=str, required=True)
-    parser.add_argument('--dataset', type=str, default='amsterdam')
+    parser.add_argument('-d', '--dataset', type=str, default='amsterdam')
     parser.add_argument('-i', '--identifier', type=str)
     parser.add_argument('-p', '--port', type=str, default='8050')
     parser.add_argument('-o', '--out-folder', type=str, default='/tmp/pyadlml/clean_device_signals')
@@ -231,11 +157,12 @@ if __name__ == '__main__':
             + "import pandas as pd\n"\
             + "from pyadlml.dataset._core.devices import get_index_matching_rows\n"
 
-    fp_code = out_folder.joinpath('code.py')
+    fp_code = out_folder.joinpath(f'clean_dev_signal_{args.device}.py')
     fp_df_devs = out_folder.joinpath('devices.csv')
 
-    with open(fp_code, mode='w') as f:
-        f.write(code_str)
+    if not fp_code.exists():
+        with open(fp_code, mode='w') as f:
+            f.write(code_str)
 
 
     # Setup data
@@ -245,26 +172,27 @@ if __name__ == '__main__':
     # Create global variables
     df_devs = data['devices']
     df_acts = data['activities']
-    write_device_df(fp_df_devs, df_devs)
+    write_devices(df_devs, fp_df_devs)
 
     assert sel_dev in df_devs[DEVICE].unique(), 'Check if device is present in dataset.'
 
     app = Dash(__name__, external_stylesheets=[dbc.themes.BOOTSTRAP])
 
-
+    #-------------------------------------------
     # Parameters to adjust 
     search_signal = [
-        (False, '6s'),
+        (True, '4s'),
+        (False, '2s'),
         (True, '1s'),
-        (False, '5s'),
-        (True, '6s')
+        (False, '6s')
     ]
-
-
 
     # The dataframe is search beforehand based on device one state (i.e. True, '1s'). Combined with a 
     # tolerance (eps_state) matches are genererated. Cross correlation is computed only around those matches.
-    matching_state_idx = 1
+    matching_state_idx = 2
+    #-------------------------------------------
+
+
     # Tolerance around matching state to be detected as a match
     eps_state = '0.2s'
     # Tolerance for cross correlation w.r.t. auto correlation to highlight as a match
@@ -333,11 +261,11 @@ if __name__ == '__main__':
 
     hits, df = create_hit_list(df_devs)
     hit_idx = hits[0]
-    sm_discrete, cross_corr, is_match = create_sig_and_corr(df, hit_idx, ss_dt_prae_match, ss_dt_post_match)
+    sm_discrete, cross_corr = create_sig_and_corr(df, hit_idx, ss_dt_prae_match, ss_dt_post_match, ss_discrete)
 
     cc_max_lst = []
     for h in hits:
-        _, cc, _ = create_sig_and_corr(df, h, ss_dt_prae_match, ss_dt_post_match)
+        _, cc = create_sig_and_corr(df, h, ss_dt_prae_match, ss_dt_post_match, ss_discrete)
         cc_max_lst.append(max(cc))
 
     ts = df_devs.loc[hit_idx, TIME]
@@ -423,13 +351,13 @@ if __name__ == '__main__':
 
         cc_max_lst = []
         for h in hits:
-            _, cc, _ = create_sig_and_corr(df, h, ss_dt_prae_match, ss_dt_post_match)
+            _, cc = create_sig_and_corr(df, h, ss_dt_prae_match, ss_dt_post_match, ss_discrete)
             cc_max_lst.append(max(cc))
 
             
         if trigger == 'cc_slider' or trigger == 'thresh_slider':
             hit_idx = hits[curr_hit]
-            sm_discrete, cross_corr, is_match = create_sig_and_corr(df, hit_idx, ss_dt_prae_match, ss_dt_post_match)
+            sm_discrete, cross_corr = create_sig_and_corr(df, hit_idx, ss_dt_prae_match, ss_dt_post_match, ss_discrete)
             fig_corr = plot_cc(sm_discrete, ss_discrete, auto_corr, cross_corr, t=int(cc_slider), eps_corr=t_slider, 
                                cc_max=cc_max_lst)
 
@@ -447,20 +375,21 @@ if __name__ == '__main__':
             code_str += f"# Hit_nr.: {curr_hit}\n"
 
             # Insert comment
+            comment = '' if comment is None else comment
             code_str += '\n'.join(['# ' + s for s in comment.split('\n')]) + '\n\n'
 
             match_idx = matches[0]
             df_tmp = df_devs[df_devs[DEVICE] == sel_dev].copy().reset_index()
             df_tmp_idx = df_tmp[df_tmp['index'] == match_idx].index[0] 
 
-            if df_tmp_idx-1 < 0:
+            if False: #df_tmp_idx-1 < 0:
                 # 1. case :|~~~e______e~~~~~    -->     Do nothing
                 #            |remove             
                 # 2. case :e~~~e______e~~~~~    -->     e______e~~~~~ 
                 #            |remove             
                 df_dev_post_idx = df_tmp.at[df_tmp_idx+1, 'index']
                 code_str += del_op2str(df_devs.loc[match_idx, :])
-                #code_str += del_op2str(df_devs.loc[df_dev_post_idx, :])
+                code_str += del_op2str(df_devs.loc[df_dev_post_idx, :])
                 print('Warning: Boundary condition removal:::: check what was done!')
             elif df_tmp_idx+1 > len(df_tmp):
                 # 1. case :______e~~~|          --> Do nothing 
@@ -497,12 +426,12 @@ if __name__ == '__main__':
             if trigger == 'prev':
                 curr_hit = max(0, curr_hit-1) 
             else:
-                curr_hit = min(curr_hit+1, len(hits))
+                curr_hit = min(curr_hit+1, len(hits)-1)
 
             comment = ''
 
             hit_idx = hits[curr_hit]
-            sm_discrete, cross_corr, is_match = create_sig_and_corr(df, hit_idx, ss_dt_prae_match, ss_dt_post_match)
+            sm_discrete, cross_corr = create_sig_and_corr(df, hit_idx, ss_dt_prae_match, ss_dt_post_match, ss_discrete)
 
             # Get area around matched state and update main window
             #ts = df.loc[df.index[df_hit_idx], TIME]
@@ -535,4 +464,4 @@ if __name__ == '__main__':
 
         return fig_main, fig_corr, curr_hit, HEADER_STR%(curr_hit, len(hits)), comment
 
-    app.run_server(debug=False)
+    app.run_server(debug=False, port=args.port, host='127.0.0.1')
